+++ src/cli/init.test.mjs
@@ -22,7 +22,9 @@ test.before(() => {
 });
 
 test.after.always(() => {
-  fs.rmSync(PROJECT_ROOT, { recursive: true });
+  if (fs.existsSync(PROJECT_ROOT)) {
+    fs.rmSync(PROJECT_ROOT, { recursive: true });
+  }
 });
 
 test("initializes 'dialect.config.json'", async (t) => {
